requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
audioop-lts==0.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.10
cors==1.0.1
cryptography==44.0.2
cvlib==0.2.7
defusedxml==0.7.1
docutils==0.21.2
dotenv==0.9.9
filelock==3.17.0
filetype==1.2.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
fsspec==2025.2.0
future==1.0.0
generativeai==0.0.1
gevent==25.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
greenlet==3.2.2
grpcio==1.70.0
grpcio-status==1.70.0
gTTS==2.5.4
httplib2==0.22.0
huggingface-hub==0.28.1
idna==3.10
imageio==2.37.0
imbalanced-learn==0.13.0
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jwt==1.3.1
Kivy==2.3.1
Kivy-Garden==0.1.5
kivy_deps.angle==0.4.0
kivy_deps.glew==0.3.1
kivy_deps.sdl2==0.8.0
kivymd==1.2.0
lazy_loader==0.4
libretranslatepy==2.1.1
lxml==5.3.1
MarkupSafe==3.0.2
MouseInfo==0.1.3
mpmath==1.3.0
mysql-connector-python==9.2.0
networkx==3.4.2
numpy==2.2.2
oauthlib==3.2.2
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.1.0
playsound==1.3.0
progressbar==2.5
proto-plus==1.26.0
protobuf==5.29.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.1
PyMsgBox==1.0.9
pyparsing==3.2.1
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.98
pytube==15.0.0
pytweening==1.2.0
pytz==2025.1
pywhatkit==5.4
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.5.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.1
setuptools==78.1.0
six==1.17.0
sklearn-compat==0.1.3
smote==0.1
soupsieve==2.6
SpeechRecognition==3.14.1
SQLAlchemy==2.0.38
standard-aifc==3.13.0
standard-chunk==3.13.0
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2025.3.30
tldextract==5.3.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.2
translate==3.6.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wikipedia==1.4.0
xgboost==2.1.3
youtube-transcript-api==0.6.3
yt-dlp==2025.2.19
zope.event==5.0
zope.interface==7.2
